Sod Blaster Tri Cities 2019 - clean code, hardware

+++ Firmware/embedded/driver/analogPinRead.cpp
@@ -16,5 +16,5 @@ analogPinRead::analogPinRead() {
  *
  */
 uint16_t analogPinRead::getBatteryVoltage() {
-	return analogRead(VOLTAGE_BATTERY_PIN);
+	return (uint16_t)analogRead(VOLTAGE_BATTERY_PIN);
 }
